Facilitate Python <3.10 compatibility
In order to enable Python <3.10 compatibility, typing-extensions needs
to be added as a dependency when `mypy-upgrade` is installed in environments
running Python <3.8 due to the `Literal` generic. Guarding the necessary
imports by system checks is required to enable isolation of the dependency to
Python <3.8 by avoiding ImportErrors.

Further, assignment expressions must be replaced with step-wise
evaluation and assignment for the same reason.

Import __future__ is required to leverage the simplified union notation added
in Python 3.10.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>