Extract error codes, description at once
In the case that the "errors" parameter of `silence_errors` is an iterator,
the iterator would have to be reset after determining if there is an
unused-ignore error. As this would have to be done both in order to extract
the error codes and the descriptions, it is better to just extract them
during the first iteration.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>