Add comparison methods for UnsilenceableRegion
By implementing custom comparison methods, the entries of
`UnsilenceableRegion`s can be typed as integers which is better for type
checkers. This also makes the implementation of the previous use case,
specifying end of lines more intuitive as -1 is commonly understood within
Python as the last in a list. Comparison then behaves as expected since
`math.inf == math.inf` evaluates to true.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>