Implement `UnsilenceableRegion.surrounds`
The only desired use of the comparison operators `__eq__` and `__lt__` was to
determine whether a given region surrounded an error. Further, the previous
implementation did not work since comparison was done between the start and
end entries not the whole object. Implementing `.surrounds` is simpler, less
code and easy to test.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>