Refactor functional tests to use env vars
The functional tests require that a particular git ref of an external
library is installed and will modify the source code of said versions.
Limiting these tests to run only within CI prevents the unwanted modification
of local development environments.

The use of environment variables to specify the mypy error report and the
package for which `mypy-upgrade` will be run lends itself to parametrization
with the CI configuration via matrices.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>