Refactor functional tests
Defining the default mypy arguments ensures that the `MYPY_UPGRADE_TARGET`
environment variable is only queried when the conditions of the CI-only tests
are met.

The class-scoped `python_path` fixture ensures that functional `mypy_upgrade`
and CLI tests obtain clean copies of the target project.

Moving mypy type checking report creation into the `mypy_report_pre` fixture
mimics an end-to-end functional test and allows for the same copy of the
target project to be linted before and after running `mypy-upgrade`.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>