Renames and refactors `message_to_error_code`
The new name `string_to_error_code` more accurately describes its function.

The refactor enables the function to return the error codes corresponding to
the longest comma-separated list of mypy error codes. This is especially
useful in the case that a string has multiple `type: ignore` phrases. The
behaviour is outlined in the example in the docstring.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>