Move `get_safe_to_silence_errors` and refactor
`get_safe_to_silence_errors` and, by extension, `UnsilenceableRegion`,
`find_unsilenceable_regions`, `find_unsilenceable_regions`,
`find_safe_end_line`, and `get_safe_to_silence_errors` are all related to
filtering `MypyError`s, so it is logical to move them to the `filter` module.

Only the line numbers of both the `start` and `end` attributes of
`UnsilenceableRegion`s are used, so implementing these attributes as integers
instead of tuples simplifies the tuple.

`_find_unsilenceable_regions` is only used by `get_safe_to_silence_errors`, so
it makes sense to make it private to the module in which it is defined.
Accepting `comments` as a sequence of strings instead of `TokenInfo` instances
enables the reuse of a list of comment strings, which can be constructed in
steps upstream of `get_safe_to_silence_errors`. It also simplifies the
second conditional check within the for loop.

Refactoring `find_safe_end_line` to return a boolean (instead of an integer)
simplifies use of its return value, since all that it is used for is to
determine if an error can be silenced. (This is brought about by the removal
of the capability to silence errors on the first line of, but still preceding,
multiline strings.) Renaming `find_safe_end_line` to `_is_safe_to_silence`
better reflects its function and the fact that it is used only by
`get_safe_to_silence_errors`.

Renaming `get_safe_to_silence_errors` to `filter_by_silenceability` better
reflects its function.

The use of keyword-only arguments is intended to improve maintainability in
the event of adding/reordering parameters.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>