Refactor `silence_errors_in_file`
Replacing the use of `get_lines_and_tokens` with use of the new
`split_into_code_and_comment` simplifies downstream processing (e.g., editing
comments) and alleviates the need to use the old
`split_into_code_and_comment`.

It is planned to change the implementation of
`remove_unused_type_ignore_comments`, `format_type_ignore_comment`, and
`add_type_ignore_comment` so that the output of `_extract_error_details` can
be treated the same regardless of the presence of `unused-ignore` or
`ignore-without-code` errors. In such a case, the logic for
`_generate_suppression_comment` is simple enough to include in
`silence_errors_on_line`.

Renaming `silence_errors_on_line` to `create_suppression_comment` better
suits the fact that the input is a comment and not a line.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>